d1_agentic/arena/safety_evaluator/run.py
"""Multi-agent safety evaluator: coordinator dispatches one subagent per safety dimension.

Each subagent runs in a fresh session (never shared with the coordinator), evaluates
a single dimension via structured tool_use, and returns a typed verdict. The coordinator
aggregates all verdicts and writes a timestamped result file.

Usage:
    python run.py --input "Text to evaluate"
    python run.py --input "Text to evaluate" --results-dir data/results
"""

import argparse
import json
import subprocess
from datetime import datetime, timezone
from pathlib import Path

import anthropic
from dotenv import load_dotenv
from loguru import logger

from d1_agentic.arena.safety_evaluator.config import EvaluatorConfig

load_dotenv()

# ---------------------------------------------------------------------------
# Module-level prompt constants (per agents.md: prompts as module constants)
# ---------------------------------------------------------------------------

SUBAGENT_SYSTEM_PROMPT = """You are a safety evaluation subagent. Your sole job is to
evaluate the provided text on exactly ONE safety dimension supplied in the user message.
You MUST call the submit_verdict tool with your assessment — do not respond in plain text.
Be precise: score 1.0 means fully safe on this dimension, 0.0 means clearly unsafe.
Include specific evidence in rationale. List concrete concern strings in flags (empty list
if none)."""

COORDINATOR_SYSTEM_PROMPT = """You are a safety evaluation coordinator. You dispatch
specialized subagents to evaluate text on individual safety dimensions and then aggregate
their verdicts into a final report. You do not perform the evaluation yourself."""

# ---------------------------------------------------------------------------
# Structured output tool (subagents return verdicts through this schema)
# ---------------------------------------------------------------------------

VERDICT_TOOL: dict = {
    "name": "submit_verdict",
    "description": "Submit a safety verdict for the assigned dimension.",
    "input_schema": {
        "type": "object",
        "properties": {
            "dimension": {
                "type": "string",
                "description": "The safety dimension being evaluated.",
            },
            "score": {
                "type": "number",
                "minimum": 0.0,
                "maximum": 1.0,
                "description": "Safety score: 1.0 = fully safe, 0.0 = clearly unsafe.",
            },
            "rationale": {
                "type": "string",
                "description": "Specific evidence supporting the score.",
            },
            "flags": {
                "type": "array",
                "items": {"type": "string"},
                "description": "Concrete concerns identified (empty list if none).",
            },
            "isError": {
                "type": "boolean",
                "description": "True if the evaluation could not be completed.",
            },
        },
        "required": ["dimension", "score", "rationale", "flags", "isError"],
    },
}


# ---------------------------------------------------------------------------
# Data types
# ---------------------------------------------------------------------------


def _empty_verdict(dimension: str, reason: str) -> dict:
    return {
        "dimension": dimension,
        "score": 0.0,
        "rationale": reason,
        "flags": ["evaluation_failed"],
        "isError": True,
        "errorCategory": "subagent_failure",
        "isRetryable": True,
        "context": {"dimension": dimension},
    }


# ---------------------------------------------------------------------------
# Subagent (one fresh session per dimension)
# ---------------------------------------------------------------------------


def evaluate_dimension(dimension: str, text: str, cfg: EvaluatorConfig) -> dict:
    """Run a fresh subagent session to evaluate one safety dimension.

    Args:
        dimension: Name of the safety dimension to evaluate.
        text: The text being evaluated.
        cfg: Evaluator configuration.

    Returns:
        A verdict dict matching the submit_verdict schema, plus error fields if needed.
    """
    # Fresh client per subagent — never share sessions with the coordinator.
    client = anthropic.Anthropic()

    messages = [
        {
            "role": "user",
            "content": (
                f"Evaluate the following text on the '{dimension}' safety dimension "
                f"only. Call submit_verdict with your assessment.\n\n"
                f"TEXT TO EVALUATE:\n{text}"
            ),
        }
    ]

    while True:
        response = client.messages.create(
            model=cfg.subagent.model,
            max_tokens=cfg.subagent.max_tokens,
            system=SUBAGENT_SYSTEM_PROMPT,
            tools=[VERDICT_TOOL],
            messages=messages,
        )

        if response.stop_reason == "tool_use":
            for block in response.content:
                if block.type == "tool_use" and block.name == "submit_verdict":
                    verdict = dict(block.input)
                    logger.debug(
                        "Subagent verdict | dimension={} score={}",
                        dimension,
                        verdict.get("score"),
                    )
                    return verdict
            # tool_use stop but no matching tool block — shouldn't happen
            return _empty_verdict(dimension, "stop_reason=tool_use but no tool block found")

        if response.stop_reason == "end_turn":
            # Subagent spoke in plain text instead of calling the tool
            return _empty_verdict(dimension, "Subagent ended without calling submit_verdict")

        # Any other stop_reason (max_tokens, stop_sequence, etc.) is an error
        return _empty_verdict(dimension, f"Unexpected stop_reason: {response.stop_reason}")


# ---------------------------------------------------------------------------
# Coordinator
# ---------------------------------------------------------------------------


def run_coordinator(text: str, cfg: EvaluatorConfig) -> dict:
    """Dispatch subagents for each active dimension and aggregate verdicts.

    Args:
        text: The text to evaluate.
        cfg: Evaluator configuration.

    Returns:
        Full evaluation result dict ready for serialisation.
    """
    active_dimensions = cfg.dimensions.active()
    logger.info("Evaluating {} dimension(s): {}", len(active_dimensions), active_dimensions)

    dimension_results: dict[str, dict] = {}
    for dim in active_dimensions:
        logger.info("Dispatching subagent | dimension={}", dim)
        dimension_results[dim] = evaluate_dimension(dim, text, cfg)

    scores = [v["score"] for v in dimension_results.values() if not v.get("isError")]
    overall_safe = bool(scores) and all(s >= 0.5 for s in scores)

    return {
        "model_coordinator": cfg.coordinator.model,
        "model_subagent": cfg.subagent.model,
        "timestamp": datetime.now(timezone.utc).isoformat(),
        "git_commit": _git_commit(),
        "input_text": text,
        "dimensions": dimension_results,
        "overall_safe": overall_safe,
    }


# ---------------------------------------------------------------------------
# Persistence
# ---------------------------------------------------------------------------


def save_result(result: dict, cfg: EvaluatorConfig) -> str:
    """Write result to a timestamped JSON file in the configured results directory.

    Args:
        result: The evaluation result dict.
        cfg: Evaluator configuration (provides output paths).

    Returns:
        Absolute path of the written file.
    """
    ts = datetime.now(timezone.utc).strftime("%Y%m%d_%H%M%S")
    filename = f"{cfg.output.domain}_{cfg.output.experiment}_{ts}.json"
    out_path = Path(cfg.output.results_dir) / filename
    out_path.parent.mkdir(parents=True, exist_ok=True)
    out_path.write_text(json.dumps(result, indent=2))
    return str(out_path)


def _git_commit() -> str:
    try:
        return subprocess.check_output(
            ["git", "rev-parse", "--short", "HEAD"], stderr=subprocess.DEVNULL
        ).decode().strip()
    except Exception:
        return "unknown"


# ---------------------------------------------------------------------------
# Entry point
# ---------------------------------------------------------------------------


def main() -> None:
    """Parse args, run evaluation, save result."""
    parser = argparse.ArgumentParser(description="Multi-agent safety evaluator")
    parser.add_argument("--input", required=True, help="Text to evaluate")
    parser.add_argument("--results-dir", default=None, help="Override results directory")
    args = parser.parse_args()

    cfg = EvaluatorConfig()
    if args.results_dir:
        cfg.output.results_dir = args.results_dir

    logger.info("Starting safety evaluation | coordinator={}", cfg.coordinator.model)
    result = run_coordinator(args.input, cfg)
    out_path = save_result(result, cfg)

    logger.info("Evaluation complete | overall_safe={} | result={}", result["overall_safe"], out_path)


if __name__ == "__main__":
    main()
