requirements.txt
anthropic>=0.49.0
python-dotenv>=1.0.0
loguru>=0.7.0
pytest>=8.0.0
pytest-mock>=3.14.0
